# docs/10-critical-remediation/api-reference-openapi.yaml
openapi: 3.1.0
info:
  title: Ubertruck MVP API Reference
  description: Complete REST API specification for Ubertruck Logistics Platform
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@ubertruck.in
  license:
    name: Proprietary
    url: https://ubertruck.in/terms
servers:
  - url: https://api.ubertruck.in/v1
    description: Production Server
  - url: https://staging-api.ubertruck.in/v1
    description: Staging Server
  - url: http://localhost:3000/v1
    description: Development Server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service communication
  schemas:
...
    Error:
      type: object
      required:
        - code
        - message
        - timestamp
        - requestId
      properties:
        code:
          type: string
          pattern: '^[A-Z]{3}_[A-Z_]+$'
          example: 'USR_NOT_FOUND'
        message:
          type: string
          example: 'User not found with the provided ID'
        details:
          type: object
          properties:
            field:
              type: string
            reason:
              type: string
            suggestion:
              type: string
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          format: uuid
        downstream:
          type: object
          properties:
            service:
              type: string
            error:
              type: string
            statusCode:
              type: integer
    Location:
      type: object
      required:
        - lat
        - lng
        - pincode
      properties:
        lat:
          type: number
          format: double
          minimum: -90
          maximum: 90
          example: 17.0005
        lng:
          type: number
          format: double
          minimum: -180
          maximum: 180
          example: 79.5926
        address:
          type: string
          maxLength: 500
        landmark:
          type: string
          maxLength: 200
        pincode:
          type: string
          pattern: '^[5][0-9]{5}$'
          example: '508001'
        city:
          type: string
          example: 'Nalgonda'
        state:
          type: string
          example: 'Telangana'
        plusCode:
          type: string
          pattern: '^[23456789CFGHJMPQRVWX]{8}\+[23456789CFGHJMPQRVWX]{2,3}$'
    ContactPerson:
      type: object
      required:
        - name
        - phoneNumber
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
        phoneNumber:
          type: string
          pattern: '^\+91[6-9]\d{9}$'
          example: '+919876543210'
        alternatePhone:
          type: string
          pattern: '^\+91[6-9]\d{9}$'

    # User Management Schemas
...
    Vehicle:
      type: object
      required:
        - registrationNumber
        - vehicleType
        - capacity
      properties:
        id:
          type: string
          format: uuid
        registrationNumber:
          type: string
          pattern: '^[A-Z]{2}[0-9]{2}[A-Z]{1,2}[0-9]{4}$'
          example: 'TG01AB1234'
        vehicleType:
          type: string
          enum: [TRUCK, MINI_TRUCK, TRAILER, CONTAINER]
        capacity:
          type: integer
          minimum: 1
          maximum: 50
          description: Capacity in tonnes
        make:
          type: string
        model:
          type: string
        year:
          type: integer
          minimum: 2000
          maximum: 2025
        fitnessValidUpto:
          type: string
          format: date
        permitValidUpto:
          type: string
          format: date
        insuranceValidUpto:
          type: string
          format: date
        gpsEnabled:
          type: boolean
        currentLocation:
          $ref: '#/components/schemas/Location'
        vahanVerified:
          type: boolean
        vahanResponse:
          type: object
          description: Vahan API verification response
    Driver:
      type: object
      required:
        - name
        - phoneNumber
        - licenseNumber
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        phoneNumber:
          type: string
          pattern: '^\+91[6-9]\d{9}$'
        licenseNumber:
          type: string
          pattern: '^[A-Z]{2}[0-9]{2}[0-9]{11}$'
        licenseValidUpto:
          type: string
          format: date
        bloodGroup:
          type: string
          enum: [A+, A-, B+, B-, AB+, AB-, O+, O-]
        emergencyContact:
          $ref: '#/components/schemas/ContactPerson'
        sarathiVerified:
          type: boolean
        sarathiResponse:
          type: object

    # Booking Management Schemas
    BookingRequest:
      type: object
      required:
        - pickupLocation
        - deliveryLocation
        - cargoDetails
        - pickupTime
      properties:
        pickupLocation:
          $ref: '#/components/schemas/Location'
        deliveryLocation:
          $ref: '#/components/schemas/Location'
        pickupTime:
          type: string
          format: date-time
        cargoDetails:
          type: object
          required:
            - type
            - weight
            - description
          properties:
            type:
              type: string
              enum: [GENERAL, FRAGILE, HAZMAT, PERISHABLE, HEAVY]
            weight:
              type: number
              minimum: 0.1
              maximum: 50
              description: Weight in tonnes
            volume:
              type: number
              description: Volume in cubic meters
            packages:
              type: integer
              minimum: 1
            description:
              type: string
              maxLength: 1000
            value:
              type: number
              description: Value for insurance in INR
            hsnCode:
              type: string
              pattern: '^\d{4,8}$'
              description: HSN code for E-Way Bill
        invoiceDetails:
          type: object
          properties:
            invoiceNumber:
              type: string
            invoiceDate:
              type: string
              format: date
            invoiceValue:
              type: number
              minimum: 0
        pickupContact:
          $ref: '#/components/schemas/ContactPerson'
        deliveryContact:
          $ref: '#/components/schemas/ContactPerson'
        specialInstructions:
          type: string
          maxLength: 500
    BookingResponse:
      type: object
      properties:
        bookingId:
          type: string
          format: uuid
        bookingNumber:
          type: string
          pattern: '^BK[0-9]{10}$'
        status:
          type: string
          enum: [CREATED, ASSIGNED, ACCEPTED, PICKUP_STARTED, IN_TRANSIT, DELIVERED, COMPLETED, CANCELLED]
        estimatedPrice:
          type: number
        assignedVehicle:
          $ref: '#/components/schemas/Vehicle'
        assignedDriver:
          $ref: '#/components/schemas/Driver'
        ewayBillNumber:
          type: string
          pattern: '^\d{12}$'
        estimatedDeliveryTime:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time

    # Payment Schemas
...
paths:
...
  /bookings:
    get:
      tags:
        - Booking Management
      summary: List user's bookings
      operationId: listBookings
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [CREATED, ASSIGNED, ACCEPTED, PICKUP_STARTED, IN_TRANSIT, DELIVERED, COMPLETED, CANCELLED]
        - in: query
          name: fromDate
          schema:
            type: string
            format: date
        - in: query
          name: toDate
          schema:
            type: string
            format: date
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Bookings retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/BookingResponse'
                  pagination:
                    type: object

    post:
      tags:
        - Booking Management
      summary: Create new booking
      operationId: createBooking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingRequest'
      responses:
        '201':
          description: Booking created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        '422':
          description: No trucks available or invalid route
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /bookings/{bookingId}/cancel:
    post:
      tags:
        - Booking Management
      summary: Cancel booking
      operationId: cancelBooking
      parameters:
        - in: path
          name: bookingId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  maxLength: 500
      responses:
        '200':
          description: Booking cancelled successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  cancellationFee:
                    type: number
                  refundAmount:
                    type: number
        '422':
          description: Booking cannot be cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /tracking/{bookingId}/pod:
    post:
      tags:
        - Tracking
      summary: Upload Proof of Delivery
      operationId: uploadPOD
      parameters:
        - in: path
          name: bookingId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - podImage
              properties:
                podImage:
                  type: string
                  format: binary
                  description: POD image (max 2MB, JPEG/PNG)
                receiverName:
                  type: string
                receiverSignature:
                  type: string
                  format: byte
                  description: Base64 encoded signature
      responses:
        '201':
          description: POD uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  podUrl:
                    type: string
                    format: uri
        '413':
          description: File too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Invalid file type or booking status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Payment Endpoints
...